+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Third party features are subject to a class's `invariant` when apply/call is used
 * A contracted feature can not be applied to objects of a different base class
 * Subcontract `extends` are now enforced
+* Updated `retry` type declaration to support accessors.
 
 ## v0.20.5